requirements.txt
# Core ML libraries
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.45.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Dataset handling
datasets>=2.14.0
huggingface_hub>=0.19.0

# Image processing
Pillow>=10.0.0

# Evaluation metrics
nltk>=3.8.0
editdistance>=0.6.0

# Streamlit app
streamlit>=1.28.0

# LaTeX rendering
matplotlib>=3.7.0

# Training utilities
wandb>=0.16.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Jupyter notebooks
jupyter>=1.0.0